[UPDATE] 멤버 도메인의 RestController 병합
